build fix

+++ src/concurrent/broadcast/broadcast_receiver.rs
@@ -56,7 +56,7 @@ impl BroadcastReceiver {
         let m_cursor = rx.buffer.get::<i64>(rx.latest_counter_index);
         let next_record = rx.cursor;
 
-        rx.record_offset = (rx.cursor & rx.mask as i64) as i32;
+        rx.record_offset = (rx.cursor & rx.mask as i64) as isize;
 
         Ok(rx)
     }
@@ -72,15 +72,15 @@ impl BroadcastReceiver {
     }
 
     pub fn type_id(&self) -> Index {
-        self.buffer.get::<i32>(record_descriptor::type_offset(self.record_offset))
+        self.buffer.get::<isize>(record_descriptor::type_offset(self.record_offset))
     }
 
     pub fn offset(&self) -> Index {
         record_descriptor::msg_offset(self.record_offset)
     }
 
-    pub fn length(&self) -> i32 {
-        self.buffer.get::<i32>(record_descriptor::length_offset(self.record_offset)) - record_descriptor::HEADER_LENGTH
+    pub fn length(&self) -> isize {
+        self.buffer.get::<isize>(record_descriptor::length_offset(self.record_offset)) - record_descriptor::HEADER_LENGTH
     }
 
     pub fn buffer(&self) -> &AtomicBuffer {
@@ -93,25 +93,25 @@ impl BroadcastReceiver {
         let mut cursor = self.next_record;
 
         if tail > cursor {
-            let mut record_offset: Index = (cursor & self.mask as i64) as i32;
+            let mut record_offset: Index = (cursor & self.mask as i64) as isize;
 
-            if !self.do_validate(cursor as i32) {
+            if !self.do_validate(cursor as isize) {
 //                self.m_lappedCount += 1;
                 cursor = self.buffer.get::<i64>(self.latest_counter_index);
-                record_offset = (cursor & self.mask as i64) as i32;
+                record_offset = (cursor & self.mask as i64) as isize;
             }
 
             self.cursor = cursor;
             self.next_record = cursor + align(
                 self.buffer.get::<i32>(
-                    record_descriptor::length_offset(record_offset)), record_descriptor::RECORD_ALIGNMENT) as i64;
+                    record_descriptor::length_offset(record_offset)) as isize, record_descriptor::RECORD_ALIGNMENT) as i64;
 
             if record_descriptor::PADDING_MSG_TYPE_ID == self.buffer.get::<i32>(record_descriptor::type_offset(record_offset))
             {
                 record_offset = 0;
                 self.cursor = self.next_record;
                 self.next_record += align(self.buffer.get::<i32>(
-                    record_descriptor::length_offset(record_offset)), record_descriptor::RECORD_ALIGNMENT) as i64;
+                    record_descriptor::length_offset(record_offset)) as isize, record_descriptor::RECORD_ALIGNMENT) as i64;
             }
 
             self.record_offset = record_offset;
@@ -123,11 +123,11 @@ impl BroadcastReceiver {
 
     pub fn validate(&self) -> bool {
         atomics::acquire();
-        self.do_validate(self.cursor as i32)
+        self.do_validate(self.cursor as isize)
     }
 
     fn do_validate(&self, cursor: Index) -> bool {
-        cursor + self.capacity > self.buffer.get_volatile::<i64>(self.tail_intent_counter_index) as i32
+        cursor + self.capacity > self.buffer.get_volatile::<i64>(self.tail_intent_counter_index) as isize
     }
 }
